+++ lib/lib_display/UDisplay/uDisplay.h
@@ -90,7 +90,8 @@ class uDisplay : public Renderer {
   uint16_t fgcol(void);
   uint16_t bgcol(void);
   int8_t color_type(void);
-  void dim(uint8_t dim);
+//   void dim(uint8_t dim);            // original version with 4 bits resolution 0..15
+  virtual void dim8(uint8_t dim, uint8_t dim_gamma);           // dimmer with 8 bits resolution, 0..255. Gamma correction must be done by caller
   uint16_t GetColorFromIndex(uint8_t index);
   void setRotation(uint8_t m);
   void fillScreen(uint16_t color);
@@ -180,9 +181,10 @@ class uDisplay : public Renderer {
    int8_t spi_clk;
    int8_t spi_mosi;
    int8_t spi_dc;
-   int8_t bpanel;
+   int8_t bpanel;          // backbanel GPIO, -1 if none
    int8_t spi_miso;
-   uint8_t dimmer;
+   uint8_t dimmer8;        // 8 bits resolution, 0..255
+   uint8_t dimmer8_gamma;  // 8 bits resolution, 0..255, gamma corrected
    SPIClass *uspi;
    uint8_t sspi;
    SPISettings spiSettings;